+++ tests/fixtures/valid_yaml/var_single_input_mixed_params.py
@@ -1,0 +1,2 @@
+def evaluate(xx, *, aa, bb):
+    return xx + aa + bb